stage_2/requirements.txt
numpy==1.26.4
pandas
scipy
tqdm
torch
torchaudio
transformers
auditok==0.2.0
whisper-timestamped==1.15.6
openai-whisper==20240930
pyannote.audio
librosa
jiwer
